requirements.txt
annotated-types==0.7.0
anyio==4.4.0
babel==2.15.0
bracex==2.4
build==1.2.1
bump-my-version==0.24.3
certifi==2024.7.4
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coverage==7.6.0
distlib==0.3.8
docutils==0.21.2
et-xmlfile==1.1.0
fastapi==0.112.0
filelock==3.15.4
ghp-import==2.1.0
griffe==0.48.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
identify==2.6.0
idna==3.7
importlib-metadata==8.2.0
iniconfig==2.0.0
jaraco-classes==3.4.0
jaraco-context==5.3.0
jaraco-functools==4.0.2
jinja2==3.1.4
keyring==25.2.1
markdown==3.6
markdown-include==0.8.1
markdown-it-py==3.0.0
markupsafe==2.1.5
mdurl==0.1.2
mergedeep==1.3.4
mkdocs==1.6.0
mkdocs-autorefs==1.0.1
mkdocs-get-deps==0.2.0
mkdocs-material==9.5.31
mkdocs-material-extensions==1.3.1
mkdocstrings==0.25.2
mkdocstrings-python==1.10.7
more-itertools==10.3.0
mypy==1.11.1
mypy-extensions==1.0.0
nh3==0.2.18
nodeenv==1.9.1
openpyxl==3.1.5
packaging==24.1
paginate==0.5.6
pathspec==0.12.1
pip==24.2
pkginfo==1.10.0
platformdirs==4.2.2
pluggy==1.5.0
pre-commit==3.8.0
prompt-toolkit==3.0.36
psycopg2-binary==2.9.9
pydantic==2.8.2
pydantic-core==2.20.1
pydantic-settings==2.4.0
pygments==2.18.0
pymdown-extensions==10.9
pyproject-hooks==1.1.0
pytest==8.3.2
pytest-cov==5.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pyyaml==6.0.1
pyyaml-env-tag==0.1
questionary==2.0.1
readme-renderer==44.0
regex==2024.7.24
requests==2.32.3
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.7.1
rich-click==1.8.3
ruff==0.5.5
six==1.16.0
sniffio==1.3.1
starlette==0.37.2
tomlkit==0.13.0
twine==5.1.1
typing-extensions==4.12.2
urllib3==2.2.2
uvicorn==0.30.5
virtualenv==20.26.3
watchdog==4.0.1
wcmatch==8.5.2
wcwidth==0.2.13
zipp==3.19.2
